Update default_resampling.py
 lower precision from float64 to float32 in the resample_data_or_seg ()

+++ umamba/nnunetv2/preprocessing/resampling/default_resampling.py
@@ -262,7 +262,7 @@ def resample_data_or_seg(data: np.ndarray, new_shape: Union[Tuple[float, ...], L
     shape = np.array(data[0].shape)
     new_shape = np.array(new_shape)
     if np.any(shape != new_shape):
-        data = data.astype(float)
+        data = data.astype(np.float32)
         if do_separate_z:
             # print("separate z, order in z is", order_z, "order inplane is", order)
             assert len(axis) == 1, "only one anisotropic axis supported"
